mqtt-tests-main/requirements.txt
cryptography==41.0.5
wsproto==1.2.0
paho-mqtt==1.6.1
websockets==12.0 
python3-dtls==1.3.0 
